+++ blocks/ad/ad.js
@@ -1,3 +1,5 @@
+import { buildAdBlock } from '../../scripts/shared.js';
+
 export default function decorate(block) {
   // Extract the unit-id from the block
   const unitIdElement = block.querySelector(
@@ -21,40 +23,9 @@ export default function decorate(block) {
   }
   const type = typeElement.textContent;
 
-  // Determine the text and class based on the type
-  let adText;
-  let adClass;
-  let height;
-  if (type === 'Advertisement') {
-    adText = 'Advertisement';
-    adClass = 'right-ad';
-    height = 'fixed-height';
-  } else if (type === 'Sponsored post') {
-    adText = 'Sponsored post';
-    adClass = 'right-sponsored';
-  } else {
-    // eslint-disable-next-line no-console
-    console.error('Unknown type in the block');
-    return;
-  }
-
   // Remove the divs containing the 'id', 'unit-id', 'type', and the type value
   const divContainers = block.querySelectorAll('div > div');
   divContainers.forEach((div) => div.remove());
 
-  // Build the ad using the extracted unit-id and determined text and class
-  const rightAdHTML = `
-    <!-- AD IMU  STARTS  -->
-
-    <div class="${adClass} ${height}">
-      <span class="ad-title">${adText}</span> <br />
-      <div id="${unitId}" class="tmsads"></div>
-    </div>
-
-    <br clear="all">
-  `;
-
-  const range = document.createRange();
-  const rightAdEl = range.createContextualFragment(rightAdHTML);
-  block.append(rightAdEl);
+  block.append(buildAdBlock(unitId, type));
 }
