+++ blocks/article-cards/article-cards.css
@@ -48,6 +48,18 @@
   height: calc(100vw * .6);
 }
 
+.article-cards .article-cards-ad {
+  background-color: #e6e6e6;
+  text-align: center;
+  padding: 0 1rem;
+  margin-top: 1rem;
+}
+
+.article-cards .sub-article .article-cards-ad {
+  width: 100%;
+  margin-bottom: auto;
+}
+
 /* lead article variant */
 .article-cards.lead-article,
 .article-cards.lead-article .sub-article .article-card:nth-child(odd) {
@@ -77,6 +89,8 @@
 
 .article-cards.lead-article .featured-article .article-card-description {
   display: block;
+  border-bottom: 2px dotted #a7a7a7;
+  padding-bottom: 2rem;
 }
 
 @media (min-width: 576px) {
